Removing json-schema-library, rewriting hooks to use the in-house variant
