+++ src/__mocks__/initialStateMock.ts
@@ -21,7 +21,6 @@ export function getInitialStateMock(customStates?: Partial<IRuntimeState>): IRun
     devTools: {
       activeTab: DevToolsTab.General,
       isOpen: false,
-      hasBeenOpen: false,
       pdfPreview: false,
       hiddenComponents: 'hide',
       layoutInspector: {
