+++ src/features/datamodel/useBindingSchema.tsx
@@ -1,40 +1,55 @@
+import { useMemo } from 'react';
+
 import type { JSONSchema7 } from 'json-schema';
 
-import { useDataModelSchema } from 'src/features/datamodel/useDataModelSchema';
+import { dotNotationToPointer } from 'src/features/datamodel/notations';
+import { lookupBindingInSchema } from 'src/features/datamodel/SimpleSchemaTraversal';
+import { useAppSelector } from 'src/hooks/useAppSelector';
+import { getCurrentDataTypeForApplication } from 'src/utils/appMetadata';
+import { getRootElementPath } from 'src/utils/schemaUtils';
 import type { IDataModelBindings } from 'src/layout/layout';
 
 type AsSchema<T> = {
   [P in keyof T]: JSONSchema7 | null;
 };
 
-// TODO: Take data into account? With proper data, we'll know exactly which schema to use in a oneOf.
+export function useCurrentDataModelName() {
+  return useAppSelector((state) =>
+    getCurrentDataTypeForApplication({
+      application: state.applicationMetadata.applicationMetadata,
+      instance: state.instanceData.instance,
+      layoutSets: state.formLayout.layoutsets,
+    }),
+  );
+}
 
 export function useBindingSchema<T extends IDataModelBindings | undefined>(bindings: T): AsSchema<T> | undefined {
-  const currentSchema = useDataModelSchema();
-  const resolvedBindings = bindings && Object.values(bindings).length ? { ...bindings } : undefined;
+  const dataModels = useAppSelector((state) => state.formDataModel.schemas);
+  const currentDataModelName = useCurrentDataModelName();
+  const currentSchema =
+    dataModels && currentDataModelName && currentDataModelName in dataModels && dataModels[currentDataModelName];
 
-  if (resolvedBindings && currentSchema) {
-    const out = {} as AsSchema<T>;
-    for (const [key, _value] of Object.entries(resolvedBindings)) {
-      const value = _value as string;
+  return useMemo(() => {
+    const resolvedBindings = bindings && Object.values(bindings).length ? { ...bindings } : undefined;
+    if (resolvedBindings && currentSchema) {
+      const rootElementPath = getRootElementPath(currentSchema);
+      const out = {} as AsSchema<T>;
+      for (const [key, _value] of Object.entries(resolvedBindings)) {
+        const value = _value as string;
+        const bindingPointer = dotNotationToPointer(value);
 
-      // Converts dot-notation to JsonPointer (including support for repeating groups)
-      const schemaPath = `/${value.replace(/\./g, '/')}`.replace(/\[(\d+)]\//g, (...a) => `/${a[1]}/`);
+        const [schema] = lookupBindingInSchema({
+          schema: currentSchema,
+          rootElementPath,
+          bindingPointer,
+        });
 
-      try {
-        const bindingSchema = currentSchema?.getSchema(schemaPath);
-        if (bindingSchema?.type === 'error') {
-          out[key] = null;
-        } else {
-          out[key] = bindingSchema;
-        }
-      } catch {
-        out[key] = null;
+        out[key] = schema || null;
       }
+
+      return out;
     }
 
-    return out;
-  }
-
-  return undefined;
+    return undefined;
+  }, [currentSchema, bindings]);
 }
