+++ src/features/devtools/data/types.ts
@@ -1,6 +1,5 @@
 export type IDevToolsState = {
   isOpen: boolean;
-  hasBeenOpen: boolean;
   pdfPreview: boolean;
   hiddenComponents: 'show' | 'disabled' | 'hide';
   activeTab: DevToolsTab;
