+++ src/index.tsx
@@ -18,7 +18,6 @@ import { App } from 'src/App';
 import { ErrorBoundary } from 'src/components/ErrorBoundary';
 import { ThemeWrapper } from 'src/components/ThemeWrapper';
 import { AppQueriesContextProvider } from 'src/contexts/appQueriesContext';
-import { DataModelSchemaContextWrapper } from 'src/features/datamodel/useDataModelSchema';
 import { DevTools } from 'src/features/devtools/DevTools';
 import * as queries from 'src/queries/queries';
 import { initSagas } from 'src/redux/sagas';
@@ -51,13 +50,11 @@ document.addEventListener('DOMContentLoaded', () => {
             <ErrorBoundary>
               <QueryClientProvider client={queryClient}>
                 <AppQueriesContextProvider {...queries}>
-                  <DataModelSchemaContextWrapper>
-                    <ExprContextWrapper>
-                      <DevTools>
-                        <App />
-                      </DevTools>
-                    </ExprContextWrapper>
-                  </DataModelSchemaContextWrapper>
+                  <ExprContextWrapper>
+                    <DevTools>
+                      <App />
+                    </DevTools>
+                  </ExprContextWrapper>
                 </AppQueriesContextProvider>
               </QueryClientProvider>
             </ErrorBoundary>
