+++ src/test/renderWithProviders.tsx
@@ -9,7 +9,6 @@ import type { PreloadedState } from 'redux';
 
 import { getInitialStateMock } from 'src/__mocks__/initialStateMock';
 import { AppQueriesContextProvider } from 'src/contexts/appQueriesContext';
-import { DataModelSchemaContextWrapper } from 'src/features/datamodel/useDataModelSchema';
 import { setupStore } from 'src/redux/store';
 import { AltinnAppTheme } from 'src/theme/altinnAppTheme';
 import { ExprContextWrapper, useResolvedNode } from 'src/utils/layout/ExprContext';
@@ -71,9 +70,7 @@ export const renderWithProviders = (
         <AppQueriesContextProvider {...mockedQueries}>
           <MuiThemeProvider theme={theme}>
             <Provider store={store}>
-              <DataModelSchemaContextWrapper>
-                <ExprContextWrapper>{children}</ExprContextWrapper>
-              </DataModelSchemaContextWrapper>
+              <ExprContextWrapper>{children}</ExprContextWrapper>
             </Provider>
           </MuiThemeProvider>
         </AppQueriesContextProvider>
